Add read/write of binary mesh files

+++ engine/renderer/VertexData.hpp
@@ -1,7 +1,8 @@
 #pragma once
+#include <cstdint>
 
 struct VertexData {
 	float* m_data = nullptr; // Continuous array of positions, uvs, normals
-	std::size_t m_numVertexData{};
-	std::size_t m_numVertices{};
+	std::uint32_t m_numVertexData{};
+	std::uint32_t m_numVertices{};
 };
